# Conflicts:
#	Examples/OverallShowcase/main.cpp
#	Source/GameDemo/main.cpp
#	Source/Runtime/include/SnowLeopardEngine/Function/Animation/Animator.h
#	Source/Runtime/include/SnowLeopardEngine/Function/Scene/Components.h
#	Source/Runtime/src/Function/Animation/Animator.cpp

+++ Source/Runtime/include/SnowLeopardEngine/Function/Animation/AnimatorController.h
@@ -16,7 +16,7 @@ namespace SnowLeopardEngine
         void DeleteAnimationClip(const Ref<AnimationClip>& animationClip);
         void SetEntryAnimationClip(const Ref<AnimationClip>& animationClip);
 
-        Ref<Transition> RegisterTransition(const Ref<AnimationClip>& sourceAnimationClip, const Ref<AnimationClip>& targetAnimationClip, int duration);
+        Ref<Transition> RegisterTransition(const Ref<AnimationClip>& sourceAnimationClip, const Ref<AnimationClip>& targetAnimationClip, float duration);
         void DeleteTransition(const Ref<AnimationClip>& sourceAnimationClip, const Ref<AnimationClip>& targetAnimationClip);
 
         void RegisterParameters(const std::string& parameterName, const std::variant<float, bool>& defaultValue);
